fixed mapping (make ids integers again)

+++ GoodsStore/GoodsStore.Domain/Entities/Coupon.cs
@@ -5,9 +5,9 @@ namespace GoodsStore.Domain.Entities
     public class Coupon
     {
         [ForeignKey("OrderDetails")] // Relations One to Zero|One (OrderDetails to Coupon)
-        public long Id { get; set; }
+        public int Id { get; set; }
 
-        public long OrderDetailsId { get; set; }
+        public int OrderDetailsId { get; set; }
 
         public virtual OrderDetails OrderDetails { get; set; }
